Commit into github repo

+++ ft_strlen.c
@@ -6,11 +6,13 @@
 /*   By: rmaes <rmaes@student.codam.nl>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/12/17 16:42:16 by rmaes             #+#    #+#             */
-/*   Updated: 2021/12/17 16:42:17 by rmaes            ###   ########.fr       */
+/*   Updated: 2022/06/13 17:19:35 by rmaes            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-unsigned long	ft_strlen(const char *s)
+#include <stddef.h>
+
+size_t	ft_strlen(const char *s)
 {
 	int	a;
